+++ tempest_fastapi_sdk/__init__.py
@@ -5,6 +5,7 @@ from tempest_fastapi_sdk.admin import (
     AdminAuthError,
     AdminModel,
     AdminSite,
+    AdminTheme,
     FieldRef,
     OrderRef,
     UserModelAuthBackend,
@@ -315,7 +316,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.71.1"
+__version__: str = "0.72.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",
@@ -344,6 +345,7 @@ __all__: list[str] = [
     "AdminAuthError",
     "AdminModel",
     "AdminSite",
+    "AdminTheme",
     "AlembicHelper",
     "AppException",
     "AsyncDatabaseManager",